# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: RareCare - Traitements API
  version: 1.0.0

servers:
  - url: http://localhost:8001/api

paths:

  /traitement:
    get:
      summary: Get all traitements
      tags:
        - Traitements
      responses:
        '200':
          description: List of traitements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Traitement'

    post:
      summary: Create a new traitement
      tags:
        - Traitements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TraitementInput'
      responses:
        '201':
          description: Traitement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Traitement'
        '422':
          description: Validation error


  /traitement/{id}:
    get:
      summary: Get a single traitement
      tags:
        - Traitements
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Traitement details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Traitement'
        '404':
          description: Traitement not found

    put:
      summary: Update a traitement
      tags:
        - Traitements
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TraitementInput'
      responses:
        '200':
          description: Traitement updated
        '404':
          description: Traitement not found

    delete:
      summary: Delete a traitement
      tags:
        - Traitements
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Traitement deleted
        '404':
          description: Traitement not found


components:
  schemas:

    Traitement:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        doctorName:
          type: string
        description:
          type: string
        created_at:
          type: string
        updated_at:
          type: string

    TraitementInput:
      type: object
      properties:
        name:
          type: string
        doctorName:
          type: string
        description:
          type: string
      required:
        - name
        - doctorName
        - description